requirements.txt
# Core
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0

# Django
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0

# Flask (if needed for any services)
Flask==3.1.1
flask-cors==6.0.1

# Server
uvicorn==0.31.1
gunicorn==23.0.0

# HTTP/API
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
requests==2.32.4

# Security
itsdangerous==2.2.0
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.1

# AI/ML
openai==1.93.0
pydantic==2.11.7
pydantic_core==2.33.2

# Image Processing
pillow==11.3.0
pytesseract==0.3.13

# Database
psycopg2-binary==2.9.10

# Utilities
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
packaging==25.0
PyYAML==6.0.2
responses==0.25.7
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
PyYAML==6.0.2
requests==2.32.4
responses==0.25.7
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
